requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
pytz==2023.3
icalendar==5.0.7
requests==2.31.0
python-dateutil==2.8.2 